Refactor namespaces to use `Extensions` and update relevant imports in source and test files
