+++ src/Algorithm/Damm/DammCalculator.cs
@@ -33,6 +33,8 @@
 
 namespace LuhnDotNet.Algorithm.Damm;
 
+using Extensions;
+
 #if NET8_0_OR_GREATER
 using System;
 #else
