+++ src/Algorithm/Luhn/LuhnCalculator.cs
@@ -38,6 +38,7 @@ using System;
 #else
 using System.Globalization;
 #endif
+using Extensions;
 using System.Diagnostics.CodeAnalysis;
