+++ src/Algorithm/Luhn/LuhnValidator.cs
@@ -33,6 +33,7 @@
 
 namespace LuhnDotNet.Algorithm.Luhn;
 
+using Extensions;
 using System;
 using System.Diagnostics.CodeAnalysis;
