+++ src/Algorithm/Mod11AscendingWeights/Mod11AscendingWeightsCalculator.cs
@@ -33,6 +33,8 @@
 
 namespace LuhnDotNet.Algorithm.Mod11AscendingWeights;
 
+using Extensions;
+
 #if NET8_0_OR_GREATER
 using System;
 #else
