+++ tests/Extension/StringExtensionTest.cs
@@ -1,6 +1,7 @@
-namespace LuhnDotNetTest;
+namespace LuhnDotNetTest.Extension;
 
 using LuhnDotNet;
+using LuhnDotNet.Extensions;
 using System;
 using System.Collections.Generic;
 using Xunit;
